Cleanup and reorganize
