+++ src/cascade/model/grids.py
@@ -3,91 +3,6 @@ from math import isclose, isinf
 GRID_SNAP_DISTANCE = 1 / 365
 
 
-class IdFactory:
-    def __init__(self):
-        self._objects = {}
-
-    def __call__(self, to_id):
-        object_hash = hash(to_id)
-        if object_hash not in self._objects:
-            if not self._objects:
-                new_id = 0
-            else:
-                new_id = max(self._objects.values()) + 1
-
-            self._objects[object_hash] = new_id
-
-        return self._objects[object_hash]
-
-
-class Prior:
-    def __init__(self, name=None):
-        self.name = name
-
-    def _parameters(self):
-        raise NotImplementedError()
-
-    def _full_parameters(self):
-        return dict(prior_name=self.name, **self._parameters())
-
-    def __hash__(self):
-        return hash(frozenset(self._full_parameters().items()))
-
-    def __eq__(self, other):
-        return isinstance(other, Prior) and self._full_parameters() == other._full_parameters()
-
-    def db_row(self, id_factory):
-        row = {
-            "prior_name": None,
-            "density": None,
-            "lower": None,
-            "upper": None,
-            "mean": None,
-            "std": None,
-            "eta": None,
-            "nu": None,
-        }.update(self._parameters())
-
-        prior_id = id_factory(self)
-
-        if row["prior_name"] is None:
-            row["prior_name"] = f"prior_{prior_id}"
-
-        return dict(prior_id=prior_id, **row)
-
-
-class UniformPrior(Prior):
-    def __init__(self, value, name=None):
-        super().__init__(name=name)
-        self.value = value
-
-    def _parameters(self):
-        return {"density": "uniform", "lower": self.value, "upper": self.value, "mean": self.value}
-
-
-class GaussianPrior(Prior):
-    def __init__(self, mean, standard_deviation, lower=float("-inf"), upper=float("inf"), name=None):
-        super().__init__(name=name)
-        if not lower <= mean <= upper:
-            raise ValueError("Bounds are inconsistent")
-        if standard_deviation < 0:
-            raise ValueError("Standard deviation must be positive")
-
-        self.lower = lower
-        self.upper = upper
-        self.mean = mean
-        self.standard_deviation = standard_deviation
-
-    def _parameters(self):
-        return {
-            "density": "gaussian",
-            "lower": self.lower,
-            "upper": self.upper,
-            "mean": self.mean,
-            "std": self.standard_deviation,
-        }
-
-
 class AgeTimeGrid:
     @classmethod
     def uniform(cls, age_start, age_end, age_step, time_start, time_end, time_step):
